FIX BUG

+++ ZQCNN/ZQ_CNN_MTCNN.h
@@ -726,7 +726,7 @@ namespace ZQ
 							block_end_h[bb] = (bh == block_num - 1) ? scoreH : ((bh + 1)*height_per_block);
 						}
 					}
-					int chunk_size = ceil(block_num / thread_num);
+					int chunk_size = ceil((float)block_num / thread_num);
 					if (thread_num <= 1)
 					{
 						for (int bb = 0; bb < block_num; bb++)
